requirements.txt
streamlit==1.37.0
openai==1.37.1
groq==0.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
streamlit-feedback==0.1.3
langchain==0.2.16
prefect==3.0.1
langchain-chroma==0.1.4
langchain-community==0.2.16
sentence-transformers==3.0.1